requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
idna==2.10
mysql-connector-python==8.0.21
protobuf==3.13.0
pycparser==2.20
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.10
